+++ lib/mix/tasks/beamlang.ex
@@ -6,6 +6,7 @@ defmodule Mix.Tasks.Beamlang do
   @moduledoc """
   Usage:
     mix beamlang path/to/file.bl [--print-tokens] [--print-ast] [--print-forms] [--emit-beam PATH] [--no-run]
+    mix beamlang --lsp
   """
 
   @spec run([binary()]) :: :ok