+++ beamlang-vscode/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "beamlang" extension will be documented in this file.
 
+## [1.1.0] - 2026-02-01
+
+### Added
+- Language server client integration for BeamLang
+- Diagnostics, hover, go-to definition, and completion
+- Configurable LSP server path and arguments
+
 ## [1.0.0] - 2026-01-31
 
 ### Added